Added google images crawling + single image training support
Added support for training the network on individual images. This does require that the network has been trained on a dataset of images before, though. So far, the network's success rate is about 80%. As for google image crawling, the program can crawl Google Images using the crawler located in the crawler folder. You can use this to lowerr your payload, so you wont have to actively search for images